requirements.txt
-i https://mirrors.aliyun.com/pypi/simple
torch
torchvision
catalyst
albumentations==1.3.0
tqdm
tensorboardX
tensorboard
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
pytorch_warmup
timm==0.6.12
transformers==4.27.1
typing_extensions==4.5.0
protobuf==3.20.1
einops
omegaconf
openai-clip
pyarrow
pytorch_metric_learning
pandas
matplotlib
diffusers